[step8] add endpoint search by brand and description

+++ application/entities/product.go
@@ -1,9 +1,9 @@
 package entities
 
 type Product struct {
-	Id          int    `bson:"id"`
-	Brand       string `bson:"brand"`
-	Description string `bson:"description"`
-	Image       string `bson:"image"`
-	Price       int    `bson:"price"`
+	Id          int     `bson:"id"`
+	Brand       string  `bson:"brand"`
+	Description string  `bson:"description"`
+	Image       string  `bson:"image"`
+	Price       float64 `bson:"price"`
 }
